@@ -1,9 +1,9 @@
 package pl.droidsonroids.casty;
 
 import android.os.Bundle;
-import android.support.annotation.CallSuper;
-import android.support.annotation.Nullable;
-import android.support.v7.app.AppCompatActivity;
+import androidx.annotation.CallSuper;
+import androidx.annotation.Nullable;
+import androidx.appcompat.app.AppCompatActivity;
 import android.view.Menu;
 import com.google.android.gms.cast.framework.media.RemoteMediaClient;
 
