@@ -1,8 +1,8 @@
 package pl.droidsonroids.casty;
 
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-import android.support.v7.app.MediaRouteButton;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.mediarouter.app.MediaRouteButton;
 import android.view.Menu;
 
 class CastyNoOp extends Casty {
