@@ -1,6 +1,6 @@
 package pl.droidsonroids.casty;
 
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 import com.google.android.gms.cast.MediaInfo;
 
 class CastyPlayerNoOp extends CastyPlayer {
