@@ -1,8 +1,8 @@
 package pl.droidsonroids.castyexample;
 
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.app.MediaRouteButton;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.mediarouter.app.MediaRouteButton;
 import android.view.Menu;
 import android.view.View;
 import android.widget.Button;
